removed old means and deviations
